+++ src/replication-protocol/index.ts
@@ -125,7 +125,8 @@ export function awaitRxStorageReplicationInSync(
 ) {
     return Promise.all([
         replicationState.streamQueue.up,
-        replicationState.streamQueue.down
+        replicationState.streamQueue.down,
+        replicationState.checkpointQueue
     ]);
 }
 
@@ -258,18 +259,14 @@ export function rxStorageInstanceToReplicationHandler<RxDocType, MasterCheckpoin
 }
 
 
-export async function cancelRxStorageReplication(
+export function cancelRxStorageReplication(
     replicationState: RxStorageInstanceReplicationState<any>
-): Promise<void> {
+) {
     replicationState.events.canceled.next(true);
-
-    await replicationState.streamQueue.down;
-    await replicationState.streamQueue.up;
-    await replicationState.checkpointQueue;
-
     replicationState.events.active.up.complete();
     replicationState.events.active.down.complete();
     replicationState.events.processed.up.complete();
     replicationState.events.processed.down.complete();
     replicationState.events.resolvedConflicts.complete();
+    replicationState.events.canceled.complete();
 }